refactor: move agent form options/submission outside of AgentConfig
